pyproject.toml
[project]
name = "chapel2"
version = "0.1.0"
description = "CadQuery project for 3D modeling with Jupyter visualization"
authors = [
    {name = "Milad"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cadquery-ocp>=7.8.0",
    "cadquery>=2.6.1,<3.0.0",
    "jupyterlab>=4.4.0,<5.0.0",
    "scipy (>=1.16.3,<2.0.0)",
    "mcp (>=1.25.0,<2.0.0)",
    "ocp-vscode (>=3.0.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
